* Removed unused header file * Fixed signed/unsigned comparison warning
svn path=/trunk/; revision=35497

+++ reactos/dll/directx/d3d9/d3d9_helpers.c
@@ -38,7 +38,7 @@ BOOL ReadRegistryValue(IN DWORD ValueType, IN LPCSTR ValueName, OUT LPBYTE DataB
     return TRUE;
 }
 
-HRESULT SafeFormatString(OUT LPSTR Buffer, IN LONG BufferSize, IN LPCSTR FormatString, ... )
+HRESULT SafeFormatString(OUT LPSTR Buffer, IN DWORD BufferSize, IN LPCSTR FormatString, ... )
 {
     DWORD BytesWritten;
     va_list vargs;
